refactor(step-1.7): rename routines/type.py -> typecls.py; avoid shadowing builtin
